Implements comics list.

+++ app/src/main/java/com/androidcourse/marvellisimo/models/comics/Price.kt
@@ -1,6 +1,6 @@
 package com.androidcourse.marvellisimo.models.comics
 
 data class Price(
-    val price: Int,
+    val price: Float,
     val type: String
 )
